Make debug tree builder macro function even less work when not debugging.
When using an instance of `BroccoliConditionalDebug` we incur a few
I/O operations even when not debugging:

* broccoli creates output path
* broccoli-conditional-debug deletes the output path
* broccoli-conditional-debug symlinks the input path to the output path

This change allows the debug tree callback to avoid even that small amount
of work. When using `BroccoliConditionalDebug.buildDebugCallback` the only
cost is the overhead is the cost of running minimatch to see if we are
debugging.