Avoid erroring if input tree is null/undefined.
Prior to this change, `null`/`undefined` was returned if the debug
flag was not set, but if the debug flag matched we would error.

It is absolutely normal for build pipelines to speculatively deal with
`null` / `undefined` input trees, and as such we should guarantee that
we will not error with or without a match.